chore: testing placed in test folder
